###TRASH_DUMP BRANCH### This kind of branches are for code that is nowhere near to be high enough in structure or quality to be pushed in any other branch
Some messing around with infinite randomly generated 2D maps.
Right now it does not work as supposed to, but i've got some basic mechanics in here.
This includes generation of grass chunks of the map putting  them in a map and render them out of this map depending on position.
Also movement is already included WASD